Ensure apt-get update is run before apt-get install
If an old image is usage, not running apt-get update may cause 404 error
when trying to install packages.
This is true even if there is apt-get update in previous RUN
instruction, because the way Docker cache thing, previous RUN of apt-get
update may use cached version (which could be weeks old).